requirements.txt
mcp
spotipy
python-dotenv
lingua-language-detector